Keep long file extensions when copying between filesystems
Copying "foo.bar1234" in the file browser produces "foo.bar" at the
destination. A second file whose name differs only after the third character of
the extension then fails with FR_EXIST, because both names collapse onto one.

FileInfo::extension means two different things depending on the source
filesystem. On FAT it is derived from the filename and truncated to three
characters by get_extension(), so the name already carries it. On D64/T64/ISO it
is CBM type metadata ("PRG", "SEQ", ...) that never appears in the name.

fcopy() called set_extension() unconditionally, which is correct for the CBM
case but rewrites the extension in the FAT case. Add ensure_extension(), which
appends only when the name does not already end in that extension, and use it
in fcopy().

Also wire the host tests into CI. They build firmware sources with the native
compiler, need no FPGA toolchain and no submodules, and run in seconds, so they
gate the much slower hardware builds. software/api/tests joins them and needed
pattern.cc on its link line, which url_encode() had made a missing symbol.

The fake filesystem that temp_auto_cleanup_test.cpp carried inline moves to
fake_filesystem.h so the new copy naming tests can mount a real source and
destination.
