+++ software/components/pattern.h
@@ -9,6 +9,7 @@ bool isEmptyString(const char *c);
 
 void set_extension(char *buffer, const char *ext, int buf_size);
 void add_extension(char *buffer, const char *ext, int buf_size);
+void ensure_extension(char *buffer, const char *ext, int buf_size);
 int  get_extension(const char *name, char *ext, bool caps=false);
 void truncate_filename(const char *orig, char *buf, int buf_size);
 int  fix_filename(char *buffer);
