+++ software/io/usb/tests/Makefile
@@ -2,6 +2,8 @@ CXX ?= g++
 CXXFLAGS := -std=c++14 -g -I. -I../../../components -I../../../system -I../../../io/stream
 KEYBOARD_TEST_FLAGS := $(CXXFLAGS) -I../../c64 -DNO_FILE_ACCESS
 
+.PHONY: all mouse keyboard keyboard-queue selection clean
+
 all: mouse keyboard keyboard-queue selection
 
 mouse:
@@ -15,3 +17,6 @@ keyboard-queue:
 
 selection:
 	$(CXX) $(CXXFLAGS) usb_hid_selection_test.cpp host_test_main.cpp -lpthread -o usbHidSelectionTest && ./usbHidSelectionTest
+
+clean:
+	rm -rf usbHidMouseTest usbHidKeyboardTest usbKeyboardQueueTest usbHidSelectionTest *.dSYM
